backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22
numpy==2.4.1
scikit-learn==1.8.0
sentence-transformers==5.2.0
langdetect==1.0.9
requests==2.32.5
